# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Services App'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'https://172.104.245.14/services-app'
paths:
...
  /api/register:
    post:
      summary: 'Register the user'
      operationId: registerTheUser
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example: {  }
                properties: {  }
      tags:
        - 'Auth APIs'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name:
                  type: string
                  description: "The vendor's first name"
                  example: et
                last_name:
                  type: string
                  description: "The vendor's last name"
                  example: amet
                email:
                  type: string
                  description: "The vendor's email"
                  example: ashly97@example.com
                phone_number:
                  type: string
                  description: "The vendor's phone number"
                  example: est
                password:
                  type: string
                  description: "The vendor's password"
                  example: 'S,etN}TTt*yX4;3Dzx'
                role:
                  type: string
                  description: "The person's role(vendor or user)"
                  example: nulla
              required:
                - first_name
                - last_name
                - email
                - phone_number
                - password
                - role
      security: []
...
  /api/reset-password:
    post:
      summary: 'Reset Password'
      operationId: resetPassword
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example: {  }
                properties: {  }
      tags:
        - 'Auth APIs'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                  description: ''
                  example: aut
                password:
                  type: string
                  description: 'New password'
                  example: '(_{Yqm>J'
                password_confirmation:
                  type: string
                  description: 'New password confirmation'
                  example: qui
              required:
                - code
                - password
                - password_confirmation
      security: []
...